1. Added Landing Form - Header 2. Added Logo to the Header 3. Added Hamburger to the Header 4. Added Search Icon to the Header 5. Installing NativeBase
