+++ android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'PieceMealV2'
+include ':react-native-vector-icons'
+project(':react-native-vector-icons').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-vector-icons/android')
 
 include ':app'
