+++ package.json
@@ -7,8 +7,10 @@
     "test": "jest"
   },
   "dependencies": {
+    "native-base": "^2.1.2",
     "react": "16.0.0-alpha.6",
-    "react-native": "0.43.4"
+    "react-native": "0.43.4",
+    "react-native-vector-icons": "^4.0.1"
   },
   "devDependencies": {
     "babel-jest": "19.0.0",
