+++ src/components/common/SearchBar.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import { View } from 'react-native';
+import { Container, Header, Item, Input, Icon, Button, Text} from 'native-base';
+
+export class SearchBar extends Component {
+    render() {
+        return (
+                  <Header searchBar rounded>
+                    <Item >
+                        <Icon name="ios-search" />
+                        <Input placeholder="Search" />
+                        <Icon name="ios-people" />
+                    </Item>
+                    <Button transparent>
+                        <Text>Search</Text>
+                    </Button>
+                  </Header>
+
+        );
+    }
+}
