+++ src/components/common/index.js
@@ -1,6 +1,7 @@
 export * from './Button';
 export * from './Card';
 export * from './CardSection';
-export * from './Header';
+export * from './MyHeader';
 export * from './Input';
 export * from './Spinner';
+export * from './SearchBar';
